pyproject.toml
[project]
name = "ai-film-studio"
version = "0.1.0"
description = "AI短剧/短片生成平台 - 基于CrewAI多智能体架构的端到端视频生成系统"
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    "crewai[tools]>=0.80.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.27.0",
    "Pillow>=10.0",
    "torch>=2.0",
    "transformers>=4.40.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "aiofiles>=23.0",
    "loguru>=0.7.0",
    "tenacity>=8.2.0",
    "edge-tts>=6.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
